+++ R/break_down.R
@@ -5,16 +5,16 @@
 #' or \code{\link{local_interactions}} for attributions with interactions.
 #'
 #' @param x an explainer created with function \code{\link[DALEX]{explain}} or a model.
-#' @param data validation dataset, will be extracted from `x` if it is an explainer.
-#' @param predict_function predict function, will be extracted from `x` if it's an explainer.
+#' @param data validation dataset, will be extracted from \code{x} if it is an explainer.
+#' @param predict_function predict function, will be extracted from \code{x} if it's an explainer.
 #' @param new_observation a new observation with columns that correspond to variables used in the model.
-#' @param keep_distributions if `TRUE`, then distribution of partial predictions is stored and can be plotted with the generic `plot()`.
-#' @param order if not `NULL`, then it will be a fixed order of variables. It can be a numeric vector or vector with names of variables.
-#' @param ... parameters passed to `local_*` functions.
+#' @param keep_distributions if \code{TRUE}, then distribution of partial predictions is stored and can be plotted with the generic \code{plot()}.
+#' @param order if not \code{NULL}, then it will be a fixed order of variables. It can be a numeric vector or vector with names of variables.
+#' @param ... parameters passed to \code{local_*} functions.
 #' @param interactions shall interactions be included?
 #' @param label name of the model. By default it is extracted from the 'class' attribute of the model.
 #'
-#' @return an object of the `break_down` class.
+#' @return an object of the \code{break_down} class.
 #'
 #' @seealso \code{\link{local_attributions}}, \code{\link{local_interactions}}
 #'